requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
arrow==1.3.0
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
binaryornot==0.4.4
boto3==1.40.49
botocore==1.40.49
celery==5.5.3
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2.3
django-allauth==64.2.1
django-redis==6.0.0
django-ses==4.4.0
django-storages==1.14.6
django-tailwind==4.2.0
elevenlabs==2.17.0
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
rich==14.2.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0